// src/frontend/webapp/modilist/src/services/swagger/api/models/create-product-input-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ModilistAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface CreateProductInputDTO
 */
export interface CreateProductInputDTO {
    /**
     * 
     * @type {string}
     * @memberof CreateProductInputDTO
     */
    sku: string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductInputDTO
     */
    name: string;
}